requirements.txt
Django==3.1.1
djangorestframework==3.11.1
flake8==3.7.9
mysqlclient==1.4.6
psycopg2==2.8.6
psycopg2-binary==2.8.6
requests==2.22.0
openpyxl==3.0.5
fpdf2==2.0.5
gunicorn==20.0.4
django-mathfilters==1.0.0
django-filter==2.3.0
python-dateutil==2.8.1
pandas==1.1.2